+++ python-automation-pro/src/utils/ibge_resolver.py
@@ -59,6 +59,7 @@ class IBGEResolver:
         "BROTAS DE MACAUBAS": "2904506", "BROTAS DE MACAÚBAS": "2904506",
         "GUARULHOS": "3518800",
         "MONTE SANTO": "2921550",
+        "SAO JOSE": "4216602", "SÃO JOSÉ": "4216602",  # São José/SC (Grande Florianópolis)
     }
 
     def __init__(self, default_uf: str = "BA", default_code: str = "2927408"):